Add daytime/nighttime weather icon support in GF PowerShell script
Introduced a new function, Test-IsDaytime, to determine if it's currently day or night based on the NWS API's isDaytime property. Updated Get-WeatherIcon to accept an isDaytime parameter, allowing for different weather icons based on the time of day. Adjusted multiple sections of the script to utilize this new functionality, enhancing the visual representation of weather conditions.